Split SongListService into Playlist and Album
